requirements.txt
fastapi
uvicorn
requests
httpx
sentence-transformers
qdrant-client
openai
pillow
python-dotenv
pandas
ultralytics
pyngrok
